budgetq/requirements.txt
pysolr
requests
django
django-haystack
